switch to SteerRequest solution
