+++ src/lemlib/motions/differential/differentialMotion.cpp
@@ -1,0 +1,7 @@
+#include "lemlib/motions/differential/differentialMotion.hpp"
+
+namespace lemlib {
+bool DifferentialMotion::isRunning() { return running; }
+
+DifferentialMotion::~DifferentialMotion() {}
+} // namespace lemlib
